split out connection.ts classes
